requirements.txt
streamlit~=1.1.0
streamlit-webrtc~=0.31.1
numpy~=1.21.3
opencv-python~=4.5.4.58
Pillow~=8.4.0